Forum unit tests

+++ forum/views.py
@@ -12,7 +12,7 @@
 
 from __future__ import unicode_literals
 
-from django.shortcuts import render
+from django.shortcuts import get_object_or_404
 from helpers.mixins import ValidateSubdomainMixin, RedirectToSignupMixin
 from student.models import Student
 from timetable.models import Semester
@@ -53,7 +53,8 @@ class ForumTranscriptView(ValidateSubdomainMixin, RedirectToSignupMixin, APIView
 
         student = Student.objects.get(user=request.user)
         semester = Semester.objects.get(name=sem_name, year=year)
-        transcript = Transcript.objects.get(owner=student, semester=semester)
+        transcript = get_object_or_404(
+            Transcript, owner=student, semester=semester)
         return Response({'transcript': TranscriptSerializer(transcript).data},
                         status=status.HTTP_200_OK)
 
@@ -67,12 +68,14 @@ class ForumTranscriptView(ValidateSubdomainMixin, RedirectToSignupMixin, APIView
 
         student = Student.objects.get(user=request.user)
         semester = Semester.objects.get(name=sem_name, year=year)
-        transcript = Transcript.objects.get(owner=Student.objects.get(jhed=request.data['jhed']),
-                                            semester=semester)
+        transcript = get_object_or_404(
+            Transcript,
+            owner=Student.objects.get(jhed=request.data['jhed']),
+            semester=semester)
 
-        if ((not student in transcript.advisors.all()) and
-                (student.jhed != transcript.owner.jhed)):
-            return Response(status=status.HTTP_401_UNAUTHORIZED)
+        if student not in transcript.advisors.all() and \
+                student.jhed != transcript.owner.jhed:
+            return Response(status=status.HTTP_403_FORBIDDEN)
 
         comment = Comment.objects.create(author=student,
                                          content=request.data['content'],
@@ -80,7 +83,7 @@ class ForumTranscriptView(ValidateSubdomainMixin, RedirectToSignupMixin, APIView
                                          transcript=transcript)
         comment.save()
 
-        return Response(status=status.HTTP_200_OK)
+        return Response(status=status.HTTP_201_CREATED)
 
     def put(self, request, sem_name, year):
         """Creates a forum transcript associated with a certain semester.
@@ -91,12 +94,15 @@ class ForumTranscriptView(ValidateSubdomainMixin, RedirectToSignupMixin, APIView
         student = Student.objects.get(user=request.user)
         semester = Semester.objects.get(name=sem_name, year=year)
 
-        transcript = Transcript.objects.create(
+        transcript = Transcript.objects.filter(
             owner=student, semester=semester)
-        transcript.save()
+        if not transcript.exists():
+            transcript = Transcript.objects.create(
+                owner=student, semester=semester)
+            transcript.save()
 
         return Response({'transcript': TranscriptSerializer(transcript).data},
-                        status=status.HTTP_200_OK)
+                        status=status.HTTP_201_CREATED)
 
     def patch(self, request, sem_name, year):
         """Adds or removes one advisor from a forum transcript.
@@ -110,13 +116,15 @@ class ForumTranscriptView(ValidateSubdomainMixin, RedirectToSignupMixin, APIView
         student = Student.objects.get(user=request.user)
         semester = Semester.objects.get(name=sem_name, year=year)
 
-        transcript = Transcript.objects.get(owner=student, semester=semester)
+        transcript = get_object_or_404(
+            Transcript, owner=student, semester=semester)
+        advisor = get_object_or_404(Student, jhed=request.data['jhed'])
         if request.data['action'] == 'add':
-            transcript.advisors.add(
-                Student.objects.get(jhed=request.data['jhed']))
+            if advisor not in transcript.advisors.all():
+                transcript.advisors.add(advisor)
         elif request.data['action'] == 'remove':
-            transcript.advisors.remove(
-                Student.objects.get(jhed=request.data['jhed']))
+            if advisor in transcript.advisors.all():
+                transcript.advisors.remove(advisor)
         transcript.save()
 
         return Response({'transcript': TranscriptSerializer(transcript).data},
@@ -128,7 +136,9 @@ class ForumTranscriptView(ValidateSubdomainMixin, RedirectToSignupMixin, APIView
 
         student = Student.objects.get(user=request.user)
         semester = Semester.objects.get(name=sem_name, year=year)
-        transcript = Transcript.objects.get(owner=student, semester=semester)
-        transcript.delete()
+        transcript = Transcript.objects.filter(
+            owner=student, semester=semester)
+        if transcript.exists():
+            transcript.delete()
 
         return Response(status=status.HTTP_204_NO_CONTENT)
